2018.04.25
- improve string table sandboxing
- move XML module to loader
