+++ openLuup/loader.lua
@@ -1,6 +1,6 @@
 local ABOUT = {
   NAME          = "openLuup.loader",
-  VERSION       = "2018.04.24",
+  VERSION       = "2018.04.25",
   DESCRIPTION   = "Loader for Device, Service, Implementation, and JSON files",
   AUTHOR        = "@akbooer",
   COPYRIGHT     = "(c) 2013-2018 AKBooer",
@@ -51,7 +51,8 @@ local ABOUT = {
 --             see: http://forum.micasaverde.com/index.php/topic,38471.0.html
 -- 2018.04.06  use scheduler.context_switch to wrap device code compilation (for sandboxing)
 -- 2018.04.22  parse_impl_xml added action field for /data_request?id=lua&DeviceNum=... 
--- 2018.04.24 replace 'extract' code in xml module (thanks again, @a-lurker)
+-- 2018.04.24  replace 'extract' code in xml module (thanks again, @a-lurker)
+-- 2018.04.25  incorporate XML module (too short, now to justify separate file)
 
 
 ------------------
@@ -74,6 +75,8 @@ local function _pristine_environment ()
     local new = shallow_copy (ENV)
     new._NAME = name                -- add environment name global
     new._G = new                    -- self reference - this IS the new global environment
+--    new.table = shallow_copy (ENV.table)    -- can sandbox individual tables like this, if necessary
+--    but string library has to be sandboxed quite differently, see scheduler sandbox function
     return new 
   end
 
@@ -82,6 +85,7 @@ local function _pristine_environment ()
   ENV.ABOUT = nil                   -- or this module's ABOUT!
   ENV.module = function () end      -- module is noop
   ENV.socket = nil                  -- somehow this got in there (from openLuup.log?)
+  ENV._G = nil                      -- each one gets its own, above
   return new_environment
 end
 
@@ -91,11 +95,127 @@ local shared_environment  = new_environment "openLuup_startup_and_scenes"
 
 ------------------
 
-local xml  = require "openLuup.xml"
 local json = require "openLuup.json"
 local vfs  = require "openLuup.virtualfilesystem"
 
 local scheduler = require "openLuup.scheduler"  -- for context_switch()
+
+------------------
+--
+-- XML module
+--
+
+-- general xml reader: this is just good enough to read device and implementation .xml files
+-- doesn't cope with XML attributes or empty elements: <tag />
+-- does cope with comments (thanks @vosmont and @a-lurker)
+--
+-- TODO: proper XML parser rather than nasty hack?
+--
+
+-- 2016.02.22  skip XML attributes but still parse element
+-- 2016.02.23  remove reader and caching, rename encode/decode 
+-- 2016.02.24  escape special characters in encode and decode
+-- 2016.04.14  @explorer expanded tags to alpha-numerics and underscores
+-- 2016.04.15  fix attribute skipping (got lost in previous edit)
+-- 2016.05.10  allow ':' as part of tag name
+
+-- 2017.03.31  make escape() and unescape() global
+
+-- 2018.04.22  remove spaces at each end of comments (part of issue highlighted by @a-lurker)
+-- see: http://forum.micasaverde.com/index.php/topic,53871.msg379551.html#msg379551
+-- and: http://forum.micasaverde.com/index.php/topic,53871.msg379790.html#msg379790
+-- 2018.04.23  ignore empty tags (remaining part of above issue?)
+-- 2018.04.24  remove extract() function completely (final part of above issue??)
+
+
+local fwd = {['<'] = "&lt;", ['>'] = "&gt;", ['"'] = "&quot;", ["'"] = "&apos;", ['&'] = "&amp;"}
+local rev = {lt = '<', gt = '>', quot = '"', apos = "'", amp = '&'}
+
+local function unescape(x)
+  return (x: gsub ("&(%w+);", rev))   -- extra parentheses to remove second return parameter
+end
+
+local function escape (x)
+  return (x: gsub ([=[[<>"'&]]=], fwd))
+end
+
+
+local function decode (info)
+  local msg
+  local xml = {}
+  -- remove such like: <!-- This is a comment -->,  thanks @vosmont
+  -- see: http://forum.micasaverde.com/index.php/topic,34572.0.html
+  if info then info = info: gsub ("%s*<!%-%-.-%-%->%s*", '') end    -- 2018.04.22 remove spaces at each end
+  --
+  local result = info
+  for a,b in (info or ''): gmatch "<([%w:_]+).->(.-)</%1>" do -- find matching opening and closing tags, ignore attributes
+    local x,y = decode (b)                                -- get the value of the contents
+    xml[a] = xml[a] or {}                                 -- if this tag doesn't exist, start a list of values
+    xml[a][#xml[a]+1] = x or y   -- add new value to the list (might be table x, or just text y)
+    result = xml
+  end 
+  if type (result) == "table" then
+    for a,b in pairs (result) do                  -- go through the contents
+      if #b == 1 then result[a] = b[1] end        -- collapse one-element lists to simple items
+    end
+  else
+    if result                     -- in case of failure, simply return whole string as 'error message'
+    and #result > 0               -- 2018.04.23 ignore empty tags
+    then
+      msg = unescape (result)
+    end
+    result = nil    -- ...and nil for xml result
+  end
+  return result, msg
+end
+
+
+local function encode (Lua, wrapper)
+  local xml = {}        -- or perhaps    {'<?xml version="1.0"?>\n'}
+  local function p(x)
+    if type (x) ~= "table" then x = {x} end
+    for _, y in ipairs (x) do xml[#xml+1] = y end
+  end
+  
+  local function value (x, name, depth)
+    local function spc ()  p ((' '):rep (2*depth)) end
+    local function atag () spc() ; p {'<', name,'>'} end
+    local function ztag () p {'</',name:match "^[^%s]+",'>\n'} end
+    local function str (x) atag() ; p(escape (tostring(x): gsub("%s+", ' '))) ; ztag() end
+    local function err (x) error ("xml: unsupported data type "..type (x)) end
+    local function tbl (x)
+      local y
+      if #x == 0 then y = {x} else y = x end
+      for i, z in ipairs (y) do
+        i = {}
+        for a in pairs (z) do i[#i+1] = a end
+        table.sort (i, function (a,b) return tostring(a) < tostring (b) end)
+        if name then atag() ; p '\n' end
+        for _,a in ipairs (i) do value(z[a], a, depth+1) end
+        if name then spc() ; ztag() end
+      end
+    end
+    
+    depth = depth or 0
+    local dispatch = {table = tbl, string = str, number = str}
+    return (dispatch [type(x)] or err) (x)
+  end
+  -- wrapper parameter allows outer level of tags (with attributes)
+  local ok, msg = pcall (value, Lua, wrapper) 
+  if ok then ok = table.concat (xml) end
+  return ok, msg
+end
+
+
+local xml = {
+   
+    escape = escape,
+    unescape = unescape,
+    
+    decode  = decode, 
+    encode  = encode,
+  }
+
 
 ------------------
 
@@ -539,4 +659,7 @@ return {
   read_impl           = read_impl,
   read_json           = read_json,
 
+  -- modules
+  xml = xml,
+  
 }
