+++ openLuup/scheduler.lua
@@ -1,6 +1,6 @@
 local ABOUT = {
   NAME          = "openLuup.scheduler",
-  VERSION       = "2018.04.23",
+  VERSION       = "2018.04.25",
   DESCRIPTION   = "openLuup job scheduler",
   AUTHOR        = "@akbooer",
   COPYRIGHT     = "(c) 2013-2018 AKBooer",
@@ -47,6 +47,7 @@ local ABOUT = {
 -- 2018.04.07  sandbox string and table system libraries
 -- 2018.04.10  add get_socket_list to methods and timenow/name to the list elements
 -- 2018.04.22  fix missing device 0 description in meta.__index:sandbox ()
+-- 2018.04.25  update sandbox function (I believe that this one actually works properly)
 
 
 local logs      = require "openLuup.logs"
@@ -185,92 +186,75 @@ local job_list = setmetatable (
 --
 -- Sandbox for system libraries
 --
-
---[[
-
-You can't do this the obvious way, because it needs to work for both this
-
-  string.foo(str, ...)
-  
-and this
-
-  str: foo (...)
-  
-This means that the actual sandbox module table MUST have a real entry,
-and not just a metatable one, as in this easy code:
-
-local function sandbox (table)
-  local meta = {}   -- index by [device][key]
-  
-  function meta:__newindex (key, value)
-    local devNo = current_device or '?'
-    meta[devNo] = meta[devNo] or {}
-    meta[devNo][key] = value
-  end
-
-  function meta:__index (key)
-    local devNo = current_device or '?'
-    local index = meta[devNo]
-    if index then return index [key] end
-  end
-  
-  return setmetatable (table, meta)
-end
-
-So it has to be done a bit differently, as below...
-
-A side-effect is that all device contexts will see the name of any additional function 
-which has been added by any device to the sandboxed module, although it may not be functional. 
-But this visibility is also true for non-sandboxed table additions, as before, so not really an issue.
-
-Note: there is STILL a potential issue here - if a function is defined twice, then it will overwrite
-the existing sandbox proxy function, since __newindex is NOT called once the table has that key.
-
-TODO: do not permit non-function definitions for sandboxed tables?
-
---]]
-
+-- Lua 5.1 strings are very special since EVERY string has a metatable with {__index = string}
+-- You can't sandbox this in the obvious way, because it needs to work for both this
+--
+--   string.foo(str, ...)
+--
+-- and this
+--
+--   str: foo (...)
+--
+-- the code below doesn't prevent modification of the original table's contents
+-- this would have to be done by an empty proxy table with its own __newindex() method
+-- other library modules can generally be sandboxed just with shallow copies
+--
+ 
 local function sandbox (tbl, name)
-  name = name or "{}"
-  local index ={}
-  local errmsg = "device %s attempt to call '%s.%s' (a nil value)"
-  local defmsg = "device %s defined '%s.%s' (a %s value)"
-  local boxmsg = "\n  [%d] %s"
-  local idxmsg = "      %s = %s"
+  
+  local devmsg = "device %s %s '%s.%s' (a %s value)"
+  local function fail(...) error(devmsg: format (...), 3) end
 
-  local meta = {__index = {}}   -- __index is only used for sandbox pretty-printing
+  name = name or "{}"
+  local lookup = {}             -- user function lookup indexed by [device][key]
+  local meta = {__index = {}}   -- used to store proxy functions for each new key
   
   function meta:__newindex(k,v)   -- only ever called if key not already defined
+  -- so this sandbox can't protect the original table keys from being changed
+  -- for that, you'd need another layer which makes a shallow copy for each user context
     
+    -- this is the proxy function which actually calls the user-defined function
     local function proxy (...) 
       local d = current_device or 0
-      local fct = (index[d] or {}) [k]
+      local fct = (lookup[d] or {}) [k]
       if not fct then
-        error (errmsg: format (d, name, k), 2)
+        fail (d, "attempted to reference", name, k, "nil")
       end
       return fct (...) 
     end
 
     local d = current_device or 0   -- k,v pairs are indexed by current device number
     local vtype = type(v)
-    _debug (defmsg: format (d, name, k, vtype))
-    index[d] = index[d] or {}
-    index[d][k] = v
-    if not self[k] then               -- only define it once (works for all device contexts)
-      rawset (self, k, (vtype == "function") and proxy or v)
+    if vtype ~= "function" then fail (d, "attempted to define", name, k, vtype) end
+    _log (devmsg: format (d, "defined", name, k, vtype), ABOUT.NAME..".sandbox")
+    lookup[d] = lookup[d] or {}
+    lookup[d][k] = v
+    if not tbl[k] then                  -- proxy only needs to be set once
+      rawset (meta.__index, k, proxy)
     end
   end
 
-  function meta.__index:sandbox ()    -- totally optional pretty-printing of sandbox contents
-    local x ={name .. ".sandbox:"}
-    for d, idx in pairs (index) do
-      local devname = (luup.devices[d] or {description = "System"}).description  -- 2018.04.22
-      x[#x+1] = boxmsg: format (d, devname: match "^%s*(.-)$")
-      for k,v in pairs (idx) do
-        x[#x+1] = idxmsg: format (k, tostring(tbl[k]))
-      end
+  function meta.__tostring ()    -- totally optional pretty-printing of sandboxed table contents
+    local boxmsg = "\n   [%d] %s"
+    local idxmsg = "        %-10s = %s"
+    local x = {name .. ".sandbox:", '', "   Private items (by device):"}
+    local empty = #x
+    local function devname (d) 
+      return ((luup.devices[d] or {}).description or "System"): match "^%s*(.+)" 
     end
-    if #x == 1 then x[2]="\n    -- none --" end
+    local function sorted(t)
+      local y = {}
+      for k,v in pairs (t) do y[#y+1] = idxmsg: format (k, tostring(v)) end
+      table.sort (y)
+      return table.concat (y, '\n')
+    end
+    for d, idx in pairs (lookup) do
+      x[#x+1] = boxmsg: format (d, devname(d))
+      x[#x+1] = sorted(idx)
+    end
+    if #x == empty then x[#x+1] ="\n        -- none --" end
+    x[#x+1] = "\n   Shared items: \n"
+    x[#x+1] = sorted (tbl) 
     x[#x+1] =''                 -- blank line at end
     return table.concat (x, '\n')
   end
@@ -278,7 +262,7 @@ local function sandbox (tbl, name)
   setmetatable (tbl, meta)
 end
 
-sandbox (table, "table")
+
 sandbox (string, "string")
 
 --
